refactor: move telegram ready code
